Upload Mechanism basically done

+++ src/main/java/com/semmelzahntiger/brainrotbackend/service/EntryService.java
@@ -1,6 +1,7 @@
 package com.semmelzahntiger.brainrotbackend.service;
 
 import com.semmelzahntiger.brainrotbackend.data.DataEntryRepository;
+import com.semmelzahntiger.brainrotbackend.data.UserRepository;
 import com.semmelzahntiger.brainrotbackend.data.entities.DataEntry;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
@@ -10,18 +11,35 @@ import java.util.UUID;
 
 @Service
 public class EntryService {
+    private final UserRepository userRepository;
     private final DataEntryRepository dataEntryRepository;
 
-    public EntryService(DataEntryRepository dataEntryRepository) {
+    public EntryService(UserRepository userRepository, DataEntryRepository dataEntryRepository) {
+        this.userRepository = userRepository;
         this.dataEntryRepository = dataEntryRepository;
     }
 
     @Transactional
-    public void updateEntries(UUID userId, List<SocialMediaResource> socialMediaResources) {
-        dataEntryRepository.deleteByUserid(userId);
+    public boolean updateEntriesOfPlatform(UUID userId, List<SocialMediaResource> socialMediaResources, SocialMediaPlatform platform) {
+        String platformName = platform.getName();
+        boolean userExists = userRepository.existsById(userId);
+        if (!userExists) {
+            return false;
+        }
+        dataEntryRepository.deleteByUseridAndPlatform(userId, platformName);
         List<DataEntry> entries = socialMediaResources.stream()
                 .map(entry -> DataEntry.fromSocialMediaResource(userId, entry)).toList();
         dataEntryRepository.saveAll(entries);
+        return true;
+    }
+    @Transactional
+    public boolean deleteEntries(UUID userId) {
+        boolean userExists = userRepository.existsById(userId);
+        if (!userExists) {
+            return false;
+        }
+        dataEntryRepository.deleteByUserid(userId);
+        return true;
     }
 
 }
